+++ firmware/controllers/algo/firing_order.h
@@ -57,6 +57,7 @@ typedef enum __attribute__ ((__packed__)) {
 	FO_1_5_4_8_6_3_7_2 = 26, // Audi 4.2 40v V8
 	FO_1_5_4_8_3_7_2_6 = 32, // Ford Voodoo
 	FO_1_8_7_3_6_5_4_2 = 28, // VH41DE (Japaneese Y32 Variant)
+	FO_1_8_6_2_7_3_4_5 = 34, // Ferrari-Maserati F136
 
 	// 9 cylinder - for instance radial :)
 	// PS: Matt says that 9cyl is actually 1-3-5-7-9-2-4-6-8 or 1-8-6-4-2-9-7-5-3 for reverse rotation
@@ -79,6 +80,6 @@ typedef enum __attribute__ ((__packed__)) {
 	// unfortunately not supported by default firmware because MAX_CYLINDER_COUNT=12 by default
 	FO_1_14_9_4_7_12_15_6_13_8_3_16_11_2_5_10 = 22, // WR16
 
-	// next value to use: 34
+	// next value to use: 35
 
 } firing_order_e;